+++ jest/jest.polyfills.js
@@ -16,6 +16,8 @@ global.Math.random = () => 0.5;
 // Suppress Console output from tested code to terminal
 console.warn = jest.fn();
 console.error = jest.fn();
+// Mock browsers sendBeacon utility
+navigator.sendBeacon = jest.fn();
 
 import { TextEncoder, TextDecoder } from 'util';