+++ packages/analytics-js-common/src/constants/destDisplayNamesToFileNamesMap.ts
@@ -148,6 +148,8 @@ import {
   TiktokAdsDirectoryName,
   ActiveCampaignDisplayName,
   ActiveCampaignDirectoryName,
+  SprigDisplayName,
+  SprigDirectoryName,
 } from './destinationNames';
 
 // The destination directory name is used as the destination SDK file name in CDN
@@ -226,6 +228,7 @@ const destDisplayNamesToFileNamesMap: Record<string, string> = {
   [LemniskDisplayName]: LemniskDirectoryName,
   [TiktokAdsDisplayName]: TiktokAdsDirectoryName,
   [ActiveCampaignDisplayName]: ActiveCampaignDirectoryName,
+  [SprigDisplayName]: SprigDirectoryName
 };
 
 export { destDisplayNamesToFileNamesMap };