+++ packages/analytics-js/__tests__/components/eventManager/RudderEventFactory.test.ts
@@ -33,6 +33,9 @@ describe('RudderEventFactory', () => {
       state.session.initialReferringDomain.value = 'https://test.com';
 
       state.consents.data.value.deniedConsentIds = ['id1', 'id2'];
+      state.consents.data.value.allowedConsentIds = ['id3', 'id4'];
+      state.consents.provider.value = 'custom';
+      state.consents.resolutionStrategy.value = undefined;
 
       state.context['ua-ch'].value = { mobile: true } as UADataValues;
       state.context.app.value = { name: 'test', version: '1.0' } as AppInfo;
@@ -41,6 +44,12 @@ describe('RudderEventFactory', () => {
       state.context.userAgent.value = 'test';
       state.context.screen.value = { width: 100, height: 100 } as ScreenInfo;
       state.context.os.value = { name: 'test', version: '1.0' } as OSInfo;
+
+      state.storage.entries.value = {
+        anonymousId: {
+          type: 'memoryStorage',
+        },
+      };
     });
   });
 
@@ -95,9 +104,6 @@ describe('RudderEventFactory', () => {
           mobile: true,
         },
         userAgent: 'test',
-        consentManagement: {
-          deniedConsentIds: ['id1', 'id2'],
-        },
         page: {
           path: '/NewPage',
           referrer: 'https://sample.com/NewPage',
@@ -180,9 +186,6 @@ describe('RudderEventFactory', () => {
           mobile: true,
         },
         userAgent: 'test',
-        consentManagement: {
-          deniedConsentIds: ['id1', 'id2'],
-        },
         page: {
           path: '/',
           referrer: '$direct',
@@ -252,9 +255,6 @@ describe('RudderEventFactory', () => {
           mobile: true,
         },
         userAgent: 'test',
-        consentManagement: {
-          deniedConsentIds: ['id1', 'id2'],
-        },
         page: {
           path: '/',
           referrer: '$direct',
@@ -337,9 +337,6 @@ describe('RudderEventFactory', () => {
           mobile: true,
         },
         userAgent: 'test',
-        consentManagement: {
-          deniedConsentIds: ['id1', 'id2'],
-        },
         page: {
           path: '/',
           referrer: '$direct',
@@ -402,9 +399,6 @@ describe('RudderEventFactory', () => {
           mobile: true,
         },
         userAgent: 'test',
-        consentManagement: {
-          deniedConsentIds: ['id1', 'id2'],
-        },
         page: {
           path: '/',
           referrer: '$direct',
@@ -439,6 +433,7 @@ describe('RudderEventFactory', () => {
 
     batch(() => {
       state.session.groupId.value = 'overridden_group_id';
+      state.consents.resolutionStrategy.value = 'or';
       state.session.groupTraits.value = {
         key4: 'value4',
         key5: 'value5',
@@ -480,9 +475,6 @@ describe('RudderEventFactory', () => {
           mobile: true,
         },
         userAgent: 'test',
-        consentManagement: {
-          deniedConsentIds: ['id1', 'id2'],
-        },
         page: {
           path: '/',
           referrer: '$direct',