+++ packages/analytics-js/src/components/core/IAnalytics.ts
@@ -1,5 +1,4 @@
 import type { IHttpClient } from '@rudderstack/analytics-js-common/types/HttpClient';
-import type { LifecycleStatus } from '@rudderstack/analytics-js-common/types/ApplicationLifecycle';
 import type { ILogger } from '@rudderstack/analytics-js-common/types/Logger';
 import type { IErrorHandler } from '@rudderstack/analytics-js-common/types/ErrorHandler';
 import type { IExternalSrcLoader } from '@rudderstack/analytics-js-common/services/ExternalSrcLoader/types';
@@ -31,7 +30,6 @@ import type { BufferQueue } from './BufferQueue';
 export interface IAnalytics {
   preloadBuffer: BufferQueue<PreloadedEventCall>;
   initialized: boolean;
-  status?: LifecycleStatus;
   httpClient: IHttpClient;
   logger: ILogger;
   errorHandler: IErrorHandler;
@@ -61,7 +59,12 @@ export interface IAnalytics {
   /**
    * Load browser polyfill if required
    */
-  prepareBrowserCapabilities(): void;
+  onMounted(): void;
+
+  /**
+   * Prepare internal services and load configuration
+   */
+  onBrowserCapabilitiesReady(): void;
 
   /**
    * Enqueue in buffer the events that were triggered pre SDK initialization
@@ -86,12 +89,12 @@ export interface IAnalytics {
   /**
    * Initialize the storage and event queue
    */
-  init(): void;
+  onPluginsReady(): void;
 
   /**
    * Load plugins
    */
-  loadPlugins(): void;
+  onConfigured(): void;
 
   /**
    * Trigger onLoaded callback if any is provided in config & emit initialised event
@@ -121,32 +124,32 @@ export interface IAnalytics {
   /**
    * To register a callback for SDK ready state
    */
-  ready(callback: ApiCallback): void;
+  ready(callback: ApiCallback, isBufferedInvocation?: boolean): void;
 
   /**
    * To record a page view event
    */
-  page(pageOptions: PageCallOptions): void;
+  page(pageOptions: PageCallOptions, isBufferedInvocation?: boolean): void;
 
   /**
    * To record a user track event
    */
-  track(trackCallOptions: TrackCallOptions): void;
+  track(trackCallOptions: TrackCallOptions, isBufferedInvocation?: boolean): void;
 
   /**
    * To record a user identification event
    */
-  identify(identifyCallOptions: IdentifyCallOptions): void;
+  identify(identifyCallOptions: IdentifyCallOptions, isBufferedInvocation?: boolean): void;
 
   /**
    * To record a user alias event
    */
-  alias(aliasCallOptions: AliasCallOptions): void;
+  alias(aliasCallOptions: AliasCallOptions, isBufferedInvocation?: boolean): void;
 
   /**
    * To record a user group event
    */
-  group(groupCallOptions: GroupCallOptions): void;
+  group(groupCallOptions: GroupCallOptions, isBufferedInvocation?: boolean): void;
 
   /**
    * To get anonymousId set in the SDK
@@ -156,12 +159,16 @@ export interface IAnalytics {
   /**
    * To set anonymousId
    */
-  setAnonymousId(anonymousId?: string, rudderAmpLinkerParam?: string): void;
+  setAnonymousId(
+    anonymousId?: string,
+    rudderAmpLinkerParam?: string,
+    isBufferedInvocation?: boolean,
+  ): void;
 
   /**
    * Clear user information, optionally anonymousId as well
    */
-  reset(resetAnonymousId?: boolean): void;
+  reset(resetAnonymousId?: boolean, isBufferedInvocation?: boolean): void;
 
   /**
    * To get userId set in the SDK
@@ -186,12 +193,12 @@ export interface IAnalytics {
   /**
    * To manually start user session in the SDK
    */
-  startSession(sessionId?: number): void;
+  startSession(sessionId?: number, isBufferedInvocation?: boolean): void;
 
   /**
    * To manually end user session in the SDK
    */
-  endSession(): void;
+  endSession(isBufferedInvocation?: boolean): void;
 
   /**
    * To fetch the current sessionId
@@ -202,7 +209,7 @@ export interface IAnalytics {
    * To record consent
    * @param options Consent API options
    */
-  consent(options?: ConsentOptions): void;
+  consent(options?: ConsentOptions, isBufferedInvocation?: boolean): void;
 
   /**
    * To set auth token