+++ packages/analytics-js/src/components/eventManager/utilities.ts
@@ -15,6 +15,7 @@ import { generateUUID } from '@rudderstack/analytics-js-common/utilities/uuId';
 import { getCurrentTimeFormatted } from '@rudderstack/analytics-js-common/utilities/timestamp';
 import { NO_STORAGE } from '@rudderstack/analytics-js-common/constants/storages';
 import { DEFAULT_INTEGRATIONS_CONFIG } from '@rudderstack/analytics-js-common/constants/integrationsConfig';
+import type { StorageType } from '@rudderstack/analytics-js-common/types/Storage';
 import { state } from '../../state';
 import {
   INVALID_CONTEXT_OBJECT_WARNING,
@@ -28,6 +29,7 @@ import {
 } from './constants';
 import { getDefaultPageProperties } from '../utilities/page';
 import { extractUTMParameters } from '../utilities/url';
+import { generateAnonymousId, isStorageTypeValidForStoringData } from '../userSessionManager/utils';
 
 /**
  * To get the page properties for context object
@@ -189,7 +191,8 @@ const getMergedContext = (
  */
 const shouldUseGlobalIntegrationsConfigInEvents = () =>
   state.loadOptions.value.useGlobalIntegrationsConfigInEvents &&
-  isObjectLiteralAndNotNull(state.nativeDestinations.loadOnlyIntegrations.value);
+  (isObjectLiteralAndNotNull(state.consents.postConsent.value?.integrations) ||
+    isObjectLiteralAndNotNull(state.nativeDestinations.loadOnlyIntegrations.value));
 
 /**
  * Updates rudder event object with data from the API options
@@ -213,7 +216,10 @@ const processOptions = (rudderEvent: RudderEvent, options?: Nullable<ApiOptions>
 const getEventIntegrationsConfig = (integrationsConfig: IntegrationOpts) => {
   let finalIntgConfig: IntegrationOpts;
   if (shouldUseGlobalIntegrationsConfigInEvents()) {
-    finalIntgConfig = state.nativeDestinations.loadOnlyIntegrations.value;
+    finalIntgConfig = clone(
+      state.consents.postConsent.value?.integrations ??
+        state.nativeDestinations.loadOnlyIntegrations.value,
+    );
   } else if (isObjectLiteralAndNotNull(integrationsConfig)) {
     finalIntgConfig = integrationsConfig;
   } else {
@@ -242,9 +248,15 @@ const getEnrichedEvent = (
       traits: clone(state.session.userTraits.value),
       sessionId: state.session.sessionInfo.value.id || undefined,
       sessionStart: state.session.sessionInfo.value.sessionStart || undefined,
-      consentManagement: {
-        deniedConsentIds: clone(state.consents.data.value.deniedConsentIds),
-      },
+      // Add 'consentManagement' only if consent management is enabled
+      ...(state.consents.enabled.value && {
+        consentManagement: {
+          deniedConsentIds: clone(state.consents.data.value.deniedConsentIds),
+          allowedConsentIds: clone(state.consents.data.value.allowedConsentIds),
+          provider: state.consents.provider.value,
+          resolutionStrategy: state.consents.resolutionStrategy.value,
+        },
+      }),
       'ua-ch': state.context['ua-ch'].value,
       app: state.context.app.value,
       library: state.context.library.value,
@@ -262,9 +274,11 @@ const getEnrichedEvent = (
     userId: rudderEvent.userId || state.session.userId.value,
   } as Partial<RudderEvent>;
 
-  if (state.storage.entries.value.anonymousId?.type === NO_STORAGE) {
+  if (
+    !isStorageTypeValidForStoringData(state.storage.entries.value.anonymousId?.type as StorageType)
+  ) {
     // Generate new anonymous id for each request
-    commonEventData.anonymousId = generateUUID();
+    commonEventData.anonymousId = generateAnonymousId();
   } else {
     // Type casting to string as the user session manager will take care of initializing the value
     commonEventData.anonymousId = state.session.anonymousId.value as string;